+++ internal/domain/ports/user.ports.go
@@ -6,7 +6,7 @@ type UserOutput interface {
 	Create(data entities.User) (entities.User, *entities.Exception)
 	Update(id string, data entities.User) (entities.User, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.User, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.User, int, *entities.Exception)
 	FindOne(id string) (entities.User, *entities.Exception)
 	FindOneByEmail(email string) (entities.User, *entities.Exception)
 }
@@ -15,6 +15,6 @@ type UserInput interface {
 	Create(data entities.User) (entities.User, *entities.Exception)
 	Update(id string, data entities.User) (entities.User, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.User, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.User, int, *entities.Exception)
 	FindOne(id string) (entities.User, *entities.Exception)
 }
